feat(tasks): die uebersicht oeffnet dieselbe aufgabe wie das aufgabenmodul
Ein Klick auf eine Aufgabe in der Uebersicht zeigte ein Kaertchen mit zwei
Knoepfen; "Bearbeiten" schickte den Nutzer ins Aufgabenmodul. Ein Aufgaben-Chip
im Kalender bot nicht einmal das an - er navigierte weg, und der Monat, den man
gerade las, war fort. Teilaufgaben, Kommentare, Dokumente und die antippbaren
Kaestchen in der Beschreibung waren von dort nicht zu erreichen.

Das wog schwerer als zwei fehlende Knoepfe, weil die Uebersicht die Ansicht ist,
in der die App im Alltag steht, und das Aufgabenmodul die, in der Aufgaben
angelegt und aufgeraeumt werden. Die haeufiger offene Ansicht konnte weniger.

Die Leseansicht wohnt deshalb nicht mehr in der Aufgabenseite, sondern in
components/task-detail.js: sie las den Betrachter aus deren `state` und lud
diese Seite nach jeder Aenderung neu - beides Wissen, das nur dort existiert.
Jetzt sagt der Aufrufer, wer schaut, was er anzeigen kann und wie er sich selbst
auffrischt. Der Kalender frischt seinen Tag auf, das Widget seine Kachel.

Was ein Feld einer Aufgabe BEDEUTET, zieht mit nach utils/task-fields.js. Diese
Regeln waren einmal abgeschrieben worden - dashboard.js fuehrte eine zweite
formatDueDate und eigene PRIORITY_LABELS.

Das Formular bleibt beim Modul und wird der Ansicht gereicht: es gehoert dorthin,
und wer es nicht mitgibt, bekommt eine Ansicht ohne Bearbeiten-Knopf statt einen,
der ins Leere fuehrt.

Closes #918
