+++ public/pages/dashboard.js
@@ -3522,62 +3522,57 @@ function initFab(signal) {
 }
 
 // --------------------------------------------------------
-// Task Quick-Action Modal
+// Eine Aufgabe aus der Übersicht öffnen (#918)
 // --------------------------------------------------------
 
-function openTaskQuickAction(taskId, taskTitle, rerender) {
-  openModal({
-    title: taskTitle,
-    size: 'sm',
-    content: `
-      <div class="modal-actions">
-        <button type="button" class="btn btn--ghost" data-action="edit">
-          <i data-lucide="edit-2" class="icon-md" aria-hidden="true"></i>
-          ${t('common.edit')}
-        </button>
-        <button type="button" class="btn btn--primary" data-action="done">
-          <i data-lucide="check-circle" class="icon-md" aria-hidden="true"></i>
-          ${t('tasks.kanbanMoveToDone')}
-        </button>
-      </div>
-    `,
-    onSave: (panel) => {
-      panel.querySelector('[data-action="done"]').addEventListener('click', async () => {
-        try {
-          await api.patch(`/tasks/${taskId}/status`, { status: 'done' });
-          closeModal({ force: true });
-          window.yuvomi?.showToast(t('tasks.swipedDoneToast'), 'success');
-          rerender();
-        } catch (err) {
-          window.yuvomi?.showToast(err.message, 'danger');
-        }
-      });
-      panel.querySelector('[data-action="edit"]').addEventListener('click', () => {
-        closeModal({ force: true });
-        window.yuvomi.navigate(`/tasks?open=${taskId}`);
-      });
-    },
-  });
+/**
+ * Dieselbe Leseansicht, die das Aufgabenmodul öffnet.
+ *
+ * Hier stand ein eigenes Kärtchen mit zwei Knöpfen: „Bearbeiten" schickte den
+ * Nutzer ins Aufgabenmodul, „Erledigt" hakte ab. Alles andere, was an einer
+ * Aufgabe hängt - Teilaufgaben, Kommentare, Dokumente, die abhakbaren Zeilen
+ * in der Beschreibung, die Fälligkeit, die Zuweisung - war von hier aus nicht
+ * zu sehen und nicht zu erreichen.
+ *
+ * Das wog schwerer, als die zwei fehlenden Knöpfe klingen: Die Übersicht ist
+ * die Ansicht, in der die App im Alltag steht. Das Aufgabenmodul ist die, in
+ * der Aufgaben angelegt und aufgeräumt werden. Die eine bot weniger an als die
+ * andere, obwohl sie öfter offen ist.
+ *
+ * Das Modul wird nachgeladen statt mitgeladen: Es zieht das Aufgabenformular
+ * mit sich, und dessen Gewicht gehört nicht in den Start der Startseite.
+ *
+ * `rerender` frischt die Kachel auf, aus der geöffnet wurde - der Nutzer
+ * bleibt, wo er war, statt im Aufgabenmodul zu landen. `user` reist mit,
+ * weil an ihm haengt, wem seine eigenen Kommentare gehoeren und wer eine
+ * gesperrte Aufgabe umschreiben darf.
+ */
+async function openTaskFromOverview(taskId, container, rerender, user) {
+  try {
+    const { openTaskById } = await import('/pages/tasks.js');
+    await openTaskById(taskId, { user, container, onChanged: rerender });
+  } catch (err) {
+    console.error('[Dashboard] Aufgabe konnte nicht geöffnet werden:', err);
+    window.yuvomi?.showToast(err.message ?? t('tasks.loadError'), 'danger');
+  }
 }
 
 // --------------------------------------------------------
 // Navigations-Links verdrahten
 // --------------------------------------------------------
 
-function wireLinks(container, rerender, { editing = false } = {}) {
+function wireLinks(container, rerender, { editing = false, user = null } = {}) {
   container.querySelectorAll('[data-route]').forEach((el) => {
     if (el.id === 'fab-main' || el.closest('#fab-actions')) return;
     if (editing && el.closest('.widget-wrapper--editing')) return;
     // Objekt-Deep-Link (Paket 2): die Cockpit-Aufgabenzeile nennt EIN Objekt,
-    // also trifft der Klick auch dieses Objekt - Quick-Action-Modal (Erledigt/
-    // Bearbeiten) wie bei den Zeilen des Tasks-Widgets, statt den Nutzer in
-    // der Aufgabenliste erneut suchen zu lassen (Critique P3). Der Titel kommt
-    // aus dem DOM: line-clamp kürzt nur visuell, textContent bleibt voll.
+    // also trifft der Klick auch dieses Objekt - die Leseansicht der Aufgabe
+    // wie bei den Zeilen des Tasks-Widgets, statt den Nutzer in der
+    // Aufgabenliste erneut suchen zu lassen (Critique P3).
     // Essen-Zeile bewusst ohne Sonderweg: /meals öffnet die aktuelle Woche und
     // scrollt den Heute-Slot selbst in den Blick (meals.js, day-header--today).
     if (!editing && el.dataset.objectKind === 'task' && el.dataset.objectId) {
-      const title = el.querySelector('.today-cockpit-card__value')?.textContent?.trim() ?? '';
-      const show = () => openTaskQuickAction(el.dataset.objectId, title, rerender);
+      const show = () => openTaskFromOverview(el.dataset.objectId, container, rerender, user);
       el.addEventListener('click', show);
       el.addEventListener('keydown', (e) => {
         if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); show(); }
@@ -3613,13 +3608,13 @@ function wireLinks(container, rerender, { editing = false } = {}) {
     });
   }
 
-  // Task-Items öffnen Quick-Action-Modal statt direkt zu navigieren
+  // Task-Items öffnen die Leseansicht statt zu navigieren (#918)
   if (editing) return;
   // Die Countdown-Zeile einer Aufgabe hängt an demselben Griff (#647): sie
   // nennt ein Objekt, also trifft der Klick dieses Objekt. Ein eigener Handler
   // daneben wäre ein zweiter Weg zur selben Handlung.
   container.querySelectorAll('.task-item[data-task-id], .countdown-item[data-task-id]').forEach((el) => {
-    const show = () => openTaskQuickAction(el.dataset.taskId, el.dataset.taskTitle, rerender);
+    const show = () => openTaskFromOverview(el.dataset.taskId, container, rerender, user);
     el.addEventListener('click', show);
     el.addEventListener('keydown', (e) => {
       if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); show(); }
@@ -4200,7 +4195,7 @@ export async function render(container, { user }) {
       </section>
       ${renderDashboardLayout(cfg, data, weather, currency, { editing: isCustomizing, visibleMealTypes, glanceHidden: !glanceVisible })}
     `);
-    wireLinks(container, rerender, { editing: isCustomizing });
+    wireLinks(container, rerender, { editing: isCustomizing, user });
     // Retry einer isolierten Widget-Fehlerkachel: da /dashboard aggregiert lädt,
     // ist „erneut versuchen" ein voller Neuaufbau (wie der Page-Level-Retry).
     container.querySelectorAll('[data-widget-retry]').forEach((btn) =>
