+++ public/sw.js
@@ -96,6 +96,7 @@ const APP_SHELL = [
   '/components/modal.js',
   '/components/photo-screensaver.js',
   '/components/quick-links-manager.js',
+  '/components/task-detail.js',
   '/components/user-multi-select.js',
   '/utils/birthday-event.js',
   '/utils/bulk-pill.js',
@@ -108,6 +109,7 @@ const APP_SHELL = [
   '/utils/dashboard-layout-hint.js',
   '/utils/dashboard-widgets.js',
   '/utils/date.js',
+  '/utils/day-label.js',
   '/utils/currency-codes.js',
   '/utils/document-preview.js',
   '/utils/event-color.js',
@@ -156,6 +158,7 @@ const APP_SHELL = [
   '/utils/swipe-row.js',
   '/utils/sync-target.js',
   '/utils/tablist.js',
+  '/utils/task-fields.js',
   '/utils/timezone.js',
   '/utils/toast-surface.js',
   '/utils/ux.js',
